requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bitarray==3.3.1
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
cytoolz==1.0.1
distro==1.9.0
duckduckgo_search==8.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
hexbytes==1.3.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonref==1.1.0
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
multidict==6.4.3
openai==1.76.0
packaging==25.0
parsimonious==0.10.0
pillow==11.2.1
primp==0.15.0
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rlp==4.1.0
six==1.17.0
smolagents==1.14.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.3
starlette==0.46.2
thirdweb-ai==0.1.16
thirdweb-mcp==0.1.16
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
web3==7.10.0
websockets==15.0.1
yarl==1.20.0
